requirements.txt
discord.py~=1.7.3
opengsq~=1.1.0
aiohttp~=3.7.4.post0